refactor(job-application): colocate each Submodel with its view and tests
The job application example kept step TEA in `step/` and step views in a
parallel `view/` tree. Project Organization says a feature folder owns
its Model, Messages, update, view, Commands, and tests.

Give every step Submodel a folder, move its view next to the TEA, and
nest list-row Submodels under `entry/`. Cover Letter keeps view in the
same file because the combined module stays small. Shared chrome
(preview, review, field helpers) stays in `src/view/`, and that folder's
index is a chrome barrel (`Field`, `Icon`, `DatePicker`). Step barrels
re-export public TEA names and view explicitly, so view-only factories
stay folder-local.

Drive each child's stories and scenes from the child update and view.
Root tests keep wizard navigation, submission, and a thin Got* fold
check per child. Shared root fixtures live in `main.fixture.ts`.

Point the generate-program skill at the new attachments folder, and bump
the foldkit-skills plugin to 0.18.19.
