+++ examples/job-application/src/step/attachments/index.ts
@@ -1,0 +1,2 @@
+export { Model, Message, init, update } from './attachments'
+export { view } from './view'
