+++ examples/job-application/src/step/coverLetter/index.ts
@@ -1,0 +1,1 @@
+export { Model, Message, init, update, view } from './coverLetter'
