+++ examples/job-application/src/step/education/entry/index.ts
@@ -1,0 +1,11 @@
+export {
+  Model,
+  Message,
+  OutMessage,
+  init,
+  update,
+  hasErrors,
+  isComplete,
+  revealErrors,
+} from './entry'
+export { type ViewInputs, view } from './view'
