+++ examples/job-application/src/step/skills/index.ts
@@ -1,2 +1,12 @@
-export * from './skills'
+export {
+  Model,
+  Message,
+  init,
+  update,
+  GenerateEntryId,
+  hasErrors,
+  isComplete,
+  revealErrors,
+} from './skills'
 export * as Entry from './entry'
+export { view } from './view'
