+++ examples/job-application/src/step/workHistory/index.ts
@@ -1,2 +1,12 @@
-export * from './workHistory'
+export {
+  Model,
+  Message,
+  init,
+  update,
+  GenerateEntryId,
+  hasErrors,
+  isComplete,
+  revealErrors,
+} from './workHistory'
 export * as Entry from './entry'
+export { view } from './view'
