--- examples/job-application/src/view/skills.ts
@@ -1,49 +1,0 @@
-import { Submodel } from 'foldkit'
-import type { Html } from 'foldkit/html'
-
-import { Button } from '@foldkit/ui'
-
-import { Skills } from '../step'
-import { Message } from '../step/skills'
-import { skillEntryView } from './skillEntry'
-
-export const skillsView = Submodel.defineView<Skills.Model, Skills.Message>(
-  (model, h): Html =>
-    h.div(
-      [h.Class('space-y-6')],
-      [
-        h.p(
-          [h.Class('text-sm text-gray-500')],
-          ['Add your technical and professional skills.'],
-        ),
-        h.div(
-          [h.Class('divide-y divide-gray-200')],
-          model.entries.map(entry =>
-            h.submodel({
-              slotId: entry.id,
-              model: entry,
-              view: skillEntryView,
-              toParentMessage: message =>
-                Message.GotEntryMessage({ entryId: entry.id, message }),
-            }),
-          ),
-        ),
-        Button.view(
-          {
-            onClick: Message.ClickedAddEntry(),
-            toView: attributes =>
-              h.button(
-                [
-                  ...attributes.button,
-                  h.Class(
-                    'w-full rounded-lg border-2 border-dashed border-gray-300 px-4 py-3 text-sm font-medium text-gray-600 hover:border-indigo-400 hover:text-indigo-600 transition cursor-pointer',
-                  ),
-                ],
-                ['+ Add Skill'],
-              ),
-          },
-          h,
-        ),
-      ],
-    ),
-)
